rock rock!

+++ Ported/AutoFarmers/Assets/Scripts/Components/ConfigurationData/FarmContent.cs
@@ -13,7 +13,7 @@ struct FarmContent : IComponentData
 	public Entity Farmer;
 	public Entity Drone;
 	public Entity Crop;
-	public Unity.Mathematics.int2 CellSize;
+	public Unity.Mathematics.float2 CellSize;
 	public uint Seed;
 	public int TeleporterCount;
 	public float Rockthreshold;
